Add a bunch of "basic" stuff.
There is still plenty that is very low hanging fruit to add (tuple,
list, Show, String, and on and on).

For what is here, great care was taken to keep the export ordering the
same as the import ordering (modulo export grouping with imports from
multiple modules).

Further, any functions that are mentioned as specializations of others
are omitted (such as `return`). Exporting and importing all of some
types and classes could go either way.

Also, note that this does not add any new dependencies or cause any
omitted functions to be errors, though we may want to do those things
in the future.