Adds compiler flag driven debugging functions
This adds a new module `Flipstone.Debug`, which conditionally
exports debugging functions with different type signatures
depending on the presence of a `DEV` compiler flag. When the flag
is present, the debugging functions have the expected behavior.
When the flag is not present, they instead will have a type
`DevFlagNotSet` that won't match and will cause the app to fail.

Co-authored-by: Ava <ava@flipstone.com>
Co-authored-by: James <james@flipstone.com>